requirements.txt
pandas
seaborn
matplotlib
requests
python-dotenv
